requirements.txt
asgiref==3.10.0
channels==4.0.0
daphne==4.0.0
Django==5.2.8
djangorestframework==3.16.1
gunicorn==23.0.0
numpy==2.3.4
pandas==2.3.3
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.7.0
